break file multipart_file_bind.go into two files
